fix(test): remove subprocess spawns causing Windows suite flakiness
The three flaky files timed out non-deterministically under full-suite load
on Windows because their fixtures spawn processes that are far more expensive
than assumed. Measured at idle on the affected host: `where <cmd>` costs
1.2-2.6s, `node -e ""` 0.7-1.9s, and `node cli.mjs <cmd> --help` ~1.6s.

runCheck spawns twice per step (commandExists, then execCommand), so the
default three-step fixture paid six spawns against a 20s budget. cli.test.mjs
spawned the CLI once per command: 26 startups measuring 21.7s against a 30s
budget, which cannot pass even on an idle machine.

Ranking each test by isolated cost against its own budget reproduces the
reported failure set exactly, in order.

Per ADR-12, remove the cost rather than raise the budgets:

- check.test.mjs / check-coverage.test.mjs: stub commandExists and execCommand
  file-wide. Neither asserts that a real process ran. Real execCommand
  behaviour stays in runner.test.mjs, which covers it in nine cases.
- cli.test.mjs: replace the 26-spawn --help loop with in-process assertions.
  That loop could not detect a flag configuration error anyway -- main()
  short-circuits --help before parseFlags builds an option table.
- cli-flags.mjs (new): extract the flag tables, loadCommandFlags and a
  buildParseOptions helper so they can be imported by a test. cli.mjs runs
  main() at import, so it cannot be imported directly. No behaviour change.

Also fixes an inverted budget: run()'s 10s child cap sat below the 6.8s idle
cost of `harness doctor`, and reported a killed child as `expected null to be
+0` rather than as a timeout. Raised to 60s and now surfaced as timedOut.

check.test.mjs test-phase time falls 53.1s -> 1.2s, check-coverage 16.8s ->
3.1s, and ~90 spawns are removed from the suite.

ADR-12's acceptance criterion now holds: three consecutive full-suite runs on
the affected Windows host, all exit 0, all reconciled, identical test-for-test
(2,329 passed / 1 skipped of 2,330 each time).
